fix: 게시글 이미지 단위 테스트 코드 임시 수정하기
